Changed the com.example.app to com.innovatelabs3.projectI2 and other minor fixes
