+++ app/build.gradle
@@ -12,11 +12,11 @@ if (localPropertiesFile.exists()) {
 }
 
 android {
-    namespace 'com.example.app'
+    namespace 'com.innovatelabs3.projectI2'
     compileSdk 34
 
     defaultConfig {
-        applicationId "com.example.app"
+        applicationId "com.innovatelabs3.projectI2"
         minSdk 24
         targetSdk 34
         versionCode 1
