+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.app">
+    package="com.innovatelabs3.projectI2">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.LlmFunctionCalling"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.innovatelabs3.projectI2.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.LlmFunctionCalling"
