--- app/src/main/java/com/example/app/MainActivity.kt
@@ -1,30 +1,0 @@
-package com.example.app
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.ui.Modifier
-import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.app.ui.screen.UserScreen
-import com.example.app.ui.theme.AppTheme
-import com.example.app.ui.viewmodel.UserViewModel
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            AppTheme {
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    val viewModel = viewModel<UserViewModel>()
-                    UserScreen(viewModel)
-                }
-            }
-        }
-    }
-}
