--- app/src/main/java/com/example/app/data/model/ChatMessage.kt
@@ -1,7 +1,0 @@
-package com.example.app.data.model
-
-data class ChatMessage(
-    val content: String,
-    val isUser: Boolean,
-    val timestamp: Long = System.currentTimeMillis()
-)
