--- app/src/main/java/com/example/app/data/model/SearchSource.kt
@@ -1,7 +1,0 @@
-package com.example.app.data.model
-
-data class SearchSource(
-    val title: String,
-    val url: String,
-    val content: String
-)
