--- app/src/main/java/com/example/app/ui/theme/Color.kt
@@ -1,38 +1,0 @@
-package com.example.app.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-// Modern color palette
-val Primary = Color(0xFF2563EB)        // Modern blue
-val PrimaryDark = Color(0xFF1E40AF)    // Darker blue
-val Secondary = Color(0xFF7C3AED)      // Purple accent
-val Background = Color(0xFFF8FAFC)     // Very light gray-blue
-val Surface = Color(0xFFFFFFFF)        // Pure white
-val TextPrimary = Color(0xFF1E293B)    // Dark blue-gray
-val TextSecondary = Color(0xFF64748B)  // Medium blue-gray
-val UserMessageBg = Primary
-val BotMessageBg = Color(0xFFFFFFFF)
-val InputBg = Color(0xFFF1F5F9)
-val Divider = Color(0xFFE2E8F0)
-val Error = Color(0xFFDC2626)
-val Success = Color(0xFF16A34A)
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
-
-// Dark Theme Colors
-val PrimaryDark_Dark = Color(0xFF60A5FA)
-val Background_Dark = Color(0xFF0F172A)
-val Surface_Dark = Color(0xFF1E293B)
-val TextPrimary_Dark = Color(0xFFF8FAFC)
-val TextSecondary_Dark = Color(0xFF94A3B8)
-val UserMessageBg_Dark = PrimaryDark_Dark
-val BotMessageBg_Dark = Color(0xFF1E293B)
-val InputBg_Dark = Color(0xFF1E293B)
-val Divider_Dark = Color(0xFF334155)
-val Warning = Color(0xFFEAB308)
