+++ app/src/main/java/com/innovatelabs3/projectI2/MainActivity.kt
@@ -1,0 +1,30 @@
+package com.innovatelabs3.projectI2
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.ui.Modifier
+import androidx.lifecycle.viewmodel.compose.viewModel
+import com.innovatelabs3.projectI2.ui.screen.UserScreen
+import com.innovatelabs3.projectI2.ui.theme.AppTheme
+import com.innovatelabs3.projectI2.ui.viewmodel.UserViewModel
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            AppTheme {
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    val viewModel = viewModel<UserViewModel>()
+                    UserScreen(viewModel)
+                }
+            }
+        }
+    }
+}
