+++ app/src/main/java/com/innovatelabs3/projectI2/data/model/ChatMessage.kt
@@ -1,0 +1,7 @@
+package com.innovatelabs3.projectI2.data.model
+
+data class ChatMessage(
+    val content: String,
+    val isUser: Boolean,
+    val timestamp: Long = System.currentTimeMillis()
+)
