+++ app/src/main/java/com/innovatelabs3/projectI2/data/model/SearchSource.kt
@@ -1,0 +1,7 @@
+package com.innovatelabs3.projectI2.data.model
+
+data class SearchSource(
+    val title: String,
+    val url: String,
+    val content: String
+)
