+++ app/src/main/java/com/innovatelabs3/projectI2/data/network/RetrofitClient.kt
@@ -1,0 +1,26 @@
+package com.innovatelabs3.projectI2.data.network
+
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import java.util.concurrent.TimeUnit
+
+object RetrofitClient {
+    private const val TAVILY_BASE_URL = "https://api.tavily.com/"
+    
+    private val okHttpClient = OkHttpClient.Builder()
+        .addInterceptor(HttpLoggingInterceptor().apply {
+            level = HttpLoggingInterceptor.Level.BODY
+        })
+        .connectTimeout(30, TimeUnit.SECONDS)
+        .readTimeout(30, TimeUnit.SECONDS)
+        .writeTimeout(30, TimeUnit.SECONDS)
+        .build()
+
+    val tavilyRetrofit: Retrofit = Retrofit.Builder()
+        .baseUrl(TAVILY_BASE_URL)
+        .client(okHttpClient)
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+}
