+++ app/src/main/java/com/innovatelabs3/projectI2/ui/theme/Type.kt
@@ -1,0 +1,37 @@
+package com.innovatelabs3.projectI2.ui.theme
+
+import androidx.compose.material3.Typography
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+import com.innovatelabs3.projectI2.R
+
+val OpenSansFont = FontFamily(
+    Font(R.font.opensans1, FontWeight.Normal)
+)
+
+val Typography = Typography(
+    bodyLarge = TextStyle(
+        fontFamily = OpenSansFont,
+        fontWeight = FontWeight.Normal,
+        fontSize = 16.sp,
+        lineHeight = 24.sp,
+        letterSpacing = 0.5.sp
+    ),
+    titleLarge = TextStyle(
+        fontFamily = OpenSansFont,
+        fontWeight = FontWeight.Normal,
+        fontSize = 22.sp,
+        lineHeight = 28.sp,
+        letterSpacing = 0.sp
+    ),
+    labelSmall = TextStyle(
+        fontFamily = OpenSansFont,
+        fontWeight = FontWeight.Normal,
+        fontSize = 11.sp,
+        lineHeight = 16.sp,
+        letterSpacing = 0.5.sp
+    )
+)
